service: Split discovery_client.py into separate sub-modules and deprecate old module
