+++ ni_measurementlink_service/_internal/service_manager.py
@@ -5,10 +5,6 @@ import grpc
 from deprecation import deprecated
 from grpc.framework.foundation import logging_pool
 
-from ni_measurementlink_service._internal.discovery_client import (
-    DiscoveryClient,
-    ServiceLocation,
-)
 from ni_measurementlink_service._internal.grpc_servicer import (
     MeasurementServiceServicerV1,
     MeasurementServiceServicerV2,
@@ -21,6 +17,7 @@ from ni_measurementlink_service._internal.stubs.ni.measurementlink.measurement.v
     measurement_service_pb2_grpc as v2_measurement_service_pb2_grpc,
 )
 from ni_measurementlink_service._loggers import ServerLogger
+from ni_measurementlink_service.discovery import DiscoveryClient, ServiceLocation
 from ni_measurementlink_service.measurement.info import MeasurementInfo, ServiceInfo
 
 _logger = logging.getLogger(__name__)
