+++ tests/acceptance/test_logging.py
@@ -6,11 +6,11 @@ import pytest
 from pytest import FixtureRequest, LogCaptureFixture
 
 from ni_measurementlink_service import session_management
-from ni_measurementlink_service._internal.discovery_client import DiscoveryClient
 from ni_measurementlink_service._internal.stubs.ni.measurementlink.measurement.v2 import (
     measurement_service_pb2 as v2_measurement_service_pb2,
     measurement_service_pb2_grpc as v2_measurement_service_pb2_grpc,
 )
+from ni_measurementlink_service.discovery import DiscoveryClient
 from ni_measurementlink_service.measurement.service import MeasurementService
 from ni_measurementlink_service.session_management import SessionManagementClient
 from tests.acceptance.test_streaming_data_measurement import (
