+++ tests/conftest.py
@@ -8,15 +8,15 @@ import pytest
 
 from ni_measurementlink_service import _featuretoggles
 from ni_measurementlink_service._featuretoggles import CodeReadiness, FeatureToggle
-from ni_measurementlink_service._internal.discovery_client import (
-    DiscoveryClient,
-    _get_registration_json_file_path,
-)
 from ni_measurementlink_service._internal.stubs.ni.measurementlink.measurement.v1 import (
     measurement_service_pb2_grpc as v1_measurement_service_pb2_grpc,
 )
 from ni_measurementlink_service._internal.stubs.ni.measurementlink.measurement.v2 import (
     measurement_service_pb2_grpc as v2_measurement_service_pb2_grpc,
+)
+from ni_measurementlink_service.discovery import DiscoveryClient
+from ni_measurementlink_service.discovery._support import (
+    _get_registration_json_file_path,
 )
 from ni_measurementlink_service.measurement.service import (
     GrpcChannelPool,
