+++ tests/unit/conftest.py
@@ -9,8 +9,8 @@ from pytest_mock import MockerFixture
 
 from ni_measurementlink_service._channelpool import GrpcChannelPool
 from ni_measurementlink_service._internal import grpc_servicer
-from ni_measurementlink_service._internal.discovery_client import DiscoveryClient
 from ni_measurementlink_service._internal.grpc_servicer import MeasurementServiceContext
+from ni_measurementlink_service.discovery import DiscoveryClient
 from ni_measurementlink_service.measurement.service import MeasurementService
 from ni_measurementlink_service.session_management import (
     MultiSessionReservation,
