+++ tests/unit/test_discovery_client.py
@@ -12,13 +12,6 @@ import pytest
 from pytest_mock import MockerFixture
 
 from ni_measurementlink_service._channelpool import GrpcChannelPool
-from ni_measurementlink_service._internal.discovery_client import (
-    DiscoveryClient,
-    ServiceLocation,
-    _get_discovery_service_address,
-    _open_key_file,
-    _start_service,
-)
 from ni_measurementlink_service._internal.stubs.ni.measurementlink.discovery.v1.discovery_service_pb2 import (
     RegisterServiceRequest,
     RegisterServiceResponse,
@@ -30,6 +23,12 @@ from ni_measurementlink_service._internal.stubs.ni.measurementlink.discovery.v1.
 )
 from ni_measurementlink_service._internal.stubs.ni.measurementlink.discovery.v1.discovery_service_pb2_grpc import (
     DiscoveryServiceStub,
+)
+from ni_measurementlink_service.discovery import DiscoveryClient, ServiceLocation
+from ni_measurementlink_service.discovery._support import (
+    _get_discovery_service_address,
+    _open_key_file,
+    _start_service,
 )
 from ni_measurementlink_service.measurement.info import MeasurementInfo, ServiceInfo
 from tests.utilities.fake_rpc_error import FakeRpcError
@@ -200,15 +199,15 @@ def test___discovery_service_not_running___get_discovery_service_address___start
     subprocess_popen_kwargs: Dict[str, Any],
 ):
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._get_key_file_path",
+        "ni_measurementlink_service.discovery._support._get_key_file_path",
         return_value=temp_discovery_key_file_path,
     )
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._service_already_running",
+        "ni_measurementlink_service.discovery._support._service_already_running",
         return_value=False,
     )
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._get_registration_json_file_path",
+        "ni_measurementlink_service.discovery._support._get_registration_json_file_path",
         return_value=temp_registration_json_file_path,
     )
     mock_popen = mocker.patch("subprocess.Popen")
@@ -233,17 +232,15 @@ def test___key_file_never_created___get_discovery_service_address___throws_timeo
     temp_registration_json_file_path: pathlib.Path,
 ):
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._get_key_file_path",
+        "ni_measurementlink_service.discovery._support._get_key_file_path",
         return_value=temp_discovery_key_file_path,
     )
+    mocker.patch("ni_measurementlink_service.discovery._support._START_SERVICE_TIMEOUT", 5.0)
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._START_SERVICE_TIMEOUT", 5.0
+        "ni_measurementlink_service.discovery._support._open_key_file", side_effect=OSError
     )
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._open_key_file", side_effect=OSError
-    )
-    mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._get_registration_json_file_path",
+        "ni_measurementlink_service.discovery._support._get_registration_json_file_path",
         return_value=temp_registration_json_file_path,
     )
     mocker.patch("subprocess.Popen")
@@ -280,7 +277,7 @@ def test___key_file_exist_after_poll___start_discovery_service___discovery_servi
     exe_file_path = temp_directory / _MOCK_REGISTRATION_FILE_CONTENT["discovery"]["path"]
 
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._open_key_file",
+        "ni_measurementlink_service.discovery._support._open_key_file",
         side_effect=[OSError, OSError, OSError, temp_discovery_key_file_path],
     )
     mock_popen = mocker.patch("subprocess.Popen")
@@ -302,11 +299,11 @@ def test___discovery_service_exe_unavailable___register_service___raises_file_no
     temp_registration_json_file_path: pathlib.Path,
 ):
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._get_key_file_path",
+        "ni_measurementlink_service.discovery._support._get_key_file_path",
         return_value=temp_discovery_key_file_path,
     )
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client._get_registration_json_file_path",
+        "ni_measurementlink_service.discovery._support._get_registration_json_file_path",
         return_value=temp_registration_json_file_path,
     )
     mocker.patch("subprocess.Popen", side_effect=FileNotFoundError)
@@ -330,7 +327,7 @@ def discovery_client(
 ) -> DiscoveryClient:
     """Create a DiscoveryClient."""
     mocker.patch(
-        "ni_measurementlink_service._internal.discovery_client.DiscoveryClient._get_stub",
+        "ni_measurementlink_service.discovery._client.DiscoveryClient._get_stub",
         return_value=discovery_service_stub,
     )
     return DiscoveryClient(grpc_channel_pool=cast(GrpcChannelPool, grpc_channel_pool))
