+++ tests/utilities/discovery_service_process.py
@@ -5,7 +5,7 @@ from typing import Optional, Type
 
 from typing_extensions import Self
 
-from ni_measurementlink_service._internal.discovery_client import (
+from ni_measurementlink_service.discovery._support import (
     _get_discovery_service_location,
     _get_key_file_path,
     _service_already_running,
