+++ tests/utilities/pin_map_client.py
@@ -7,11 +7,11 @@ from typing import Optional, Union
 import grpc
 
 from ni_measurementlink_service._channelpool import GrpcChannelPool
-from ni_measurementlink_service._internal.discovery_client import DiscoveryClient
 from ni_measurementlink_service._internal.stubs.ni.measurementlink.pinmap.v1 import (
     pin_map_service_pb2,
     pin_map_service_pb2_grpc,
 )
+from ni_measurementlink_service.discovery import DiscoveryClient
 
 _logger = logging.getLogger(__name__)
